remove core/print.go
