+++ cmd/registry/cmd/compute/complexity/complexity.go
@@ -27,6 +27,7 @@ import (
 	"github.com/apigee/registry/pkg/visitor"
 	"github.com/apigee/registry/rpc"
 	"github.com/spf13/cobra"
+	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
 
 	discovery "github.com/google/gnostic/discovery"
@@ -169,7 +170,7 @@ func (task *computeComplexityTask) Run(ctx context.Context) error {
 	}
 
 	if task.dryRun {
-		core.PrintMessage(complexity)
+		fmt.Println(protojson.Format(complexity))
 		return nil
 	}
 	subject := task.specName
