+++ cmd/registry/cmd/compute/lint/lint.go
@@ -26,6 +26,7 @@ import (
 	"github.com/apigee/registry/pkg/visitor"
 	"github.com/apigee/registry/rpc"
 	"github.com/spf13/cobra"
+	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -147,7 +148,7 @@ func (task *computeLintTask) Run(ctx context.Context) error {
 	}
 
 	if task.dryRun {
-		core.PrintMessage(lint)
+		fmt.Println(protojson.Format((lint)))
 		return nil
 	}
