+++ cmd/registry/cmd/vocabulary/unique.go
@@ -15,14 +15,15 @@
 package vocabulary
 
 import (
+	"fmt"
 	"path/filepath"
 	"strings"
 
-	"github.com/apigee/registry/cmd/registry/core"
 	"github.com/apigee/registry/log"
 	"github.com/apigee/registry/pkg/connection"
 	"github.com/google/gnostic/metrics/vocabulary"
 	"github.com/spf13/cobra"
+	"google.golang.org/protobuf/encoding/protojson"
 )
 
 func uniqueCommand() *cobra.Command {
@@ -54,7 +55,7 @@ func uniqueCommand() *cobra.Command {
 					setVocabularyToArtifact(ctx, client, unique, outputArtifactName)
 				}
 			} else {
-				core.PrintMessage(list)
+				fmt.Println(protojson.Format((list)))
 			}
 		},
 	}
