+++ cmd/registry/conformance/conformance-task.go
@@ -29,6 +29,7 @@ import (
 	"github.com/apigee/registry/pkg/connection"
 	"github.com/apigee/registry/pkg/names"
 	"github.com/apigee/registry/rpc"
+	"google.golang.org/protobuf/encoding/protojson"
 	"google.golang.org/protobuf/proto"
 )
 
@@ -136,7 +137,7 @@ func (task *ComputeConformanceTask) Run(ctx context.Context) error {
 	}
 
 	if task.DryRun {
-		core.PrintMessage(conformanceReport)
+		fmt.Println(protojson.Format((conformanceReport)))
 		return nil
 	}
 	return task.storeConformanceReport(ctx, conformanceReport)
